moved out QueryBuilder into its own file within the internal library query
