+++ lib/src/query/index.dart
@@ -1,0 +1,1 @@
+export "query.dart";
